+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Support for CRSH greeting message format 2.
 - Sending of config errors/warnings to ccache.
 
+### Fixed
+
+- Unnecessary serialization of storage client connections.
+
 ### Changed
 
 - Improve logging of failures.